rm int()

+++ sb_discriminator/train.py
@@ -80,13 +80,13 @@ if __name__ == "__main__":
             f"Loaded model from {args.load_model if args.load_model else args.eval_model} at epoch {loaded_epoch} with best validation loss {best_vloss} and best validation accuracy {best_vaccuracy}"
         )
 
-    train_batch_prints = int(len(training_loader) / args.num_prints)
-    num_train_batches = int(len(training_loader))
+    train_batch_prints = len(training_loader) / args.num_prints
+    num_train_batches = len(training_loader)
     print("train_batch_prints", train_batch_prints)
 
     if not args.eval_model:
-        val_batch_prints = int(len(val_loader) / args.num_prints)
-        num_val_batches = int(len(val_loader))
+        val_batch_prints = len(val_loader) / args.num_prints
+        num_val_batches = len(val_loader)
         print("val_batch_prints", val_batch_prints)
 
         for epoch in range(args.epochs):
@@ -204,8 +204,8 @@ if __name__ == "__main__":
         logger.info("Evaluating best model on test and train dataset")
         print("================================")
 
-        test_batch_prints = int(len(test_loader) / args.num_prints)
-        num_test_batches = int(len(test_loader))
+        test_batch_prints = len(test_loader) / args.num_prints
+        num_test_batches = len(test_loader)
 
         print("test_batch_prints", test_batch_prints)
